backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.2.0
torchvision==0.17.0
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
pydantic==2.5.0
pytorch-grad-cam==1.5.2
python-dotenv==1.0.0
tensorflow==2.14.0